Added map fragment and reorganized/renamed fragments
