+++ Moodlet/app/src/main/java/com/cmput3owo1/moodlet/activities/MainActivity.java
@@ -1,7 +1,8 @@
-package com.cmput3owo1.moodlet;
+package com.cmput3owo1.moodlet.activities;
 
 import android.os.Bundle;
 
+import com.cmput3owo1.moodlet.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 import androidx.appcompat.app.AppCompatActivity;
@@ -20,7 +21,8 @@ public class MainActivity extends AppCompatActivity {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications)
+                R.id.navigation_mood_history, R.id.navigation_feed, R.id.navigation_search,
+                R.id.navigation_map)
                 .build();
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
