+++ Moodlet/app/src/main/java/com/cmput3owo1/moodlet/fragments/FeedFragment.java
@@ -1,0 +1,19 @@
+package com.cmput3owo1.moodlet.fragments;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+
+import com.cmput3owo1.moodlet.R;
+
+public class FeedFragment extends Fragment {
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_feed, container, false);
+    }
+}
