+++ Moodlet/app/src/main/res/menu/bottom_nav_menu.xml
@@ -2,18 +2,23 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/navigation_home"
+        android:id="@+id/navigation_mood_history"
         android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
+        android:title="@string/title_history" />
 
     <item
-        android:id="@+id/navigation_dashboard"
+        android:id="@+id/navigation_feed"
         android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
+        android:title="@string/title_friend_activity" />
 
     <item
-        android:id="@+id/navigation_notifications"
+        android:id="@+id/navigation_search"
         android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
+        android:title="@string/title_discover" />
+
+    <item
+        android:id="@+id/navigation_map"
+        android:icon="@drawable/ic_notifications_black_24dp"
+        android:title="@string/title_map" />
 
 </menu>
