+++ Moodlet/app/src/main/res/navigation/mobile_navigation.xml
@@ -3,23 +3,29 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_home">
+    app:startDestination="@+id/navigation_mood_history">
 
     <fragment
-        android:id="@+id/navigation_home"
-        android:name="com.cmput3owo1.moodlet.ui.home.HomeFragment"
-        android:label="@string/title_home"
-        tools:layout="@layout/fragment_home" />
+        android:id="@+id/navigation_mood_history"
+        android:name="com.cmput3owo1.moodlet.fragments.MoodHistoryFragment"
+        android:label="@string/title_history"
+        tools:layout="@layout/fragment_mood_history" />
 
     <fragment
-        android:id="@+id/navigation_dashboard"
-        android:name="com.cmput3owo1.moodlet.ui.dashboard.DashboardFragment"
-        android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" />
+        android:id="@+id/navigation_feed"
+        android:name="com.cmput3owo1.moodlet.fragments.FeedFragment"
+        android:label="@string/title_friend_activity"
+        tools:layout="@layout/fragment_feed" />
 
     <fragment
-        android:id="@+id/navigation_notifications"
-        android:name="com.cmput3owo1.moodlet.ui.notifications.NotificationsFragment"
-        android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        android:id="@+id/navigation_search"
+        android:name="com.cmput3owo1.moodlet.fragments.SearchFragment"
+        android:label="@string/title_discover"
+        tools:layout="@layout/fragment_search" />
+
+    <fragment
+        android:id="@+id/navigation_map"
+        android:name="com.cmput3owo1.moodlet.fragments.MapFragment"
+        android:label="@string/title_map"
+        tools:layout="@layout/fragment_map" />
 </navigation>
