+++ Moodlet/app/src/main/res/values/strings.xml
@@ -1,6 +1,7 @@
 <resources>
     <string name="app_name">Moodlet</string>
-    <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications</string>
+    <string name="title_history">History</string>
+    <string name="title_friend_activity">Friend Activity</string>
+    <string name="title_discover">Discover</string>
+    <string name="title_map">Map</string>
 </resources>
